refactor(explain): extract plan analysis and views into @tabularis/explain
The plan analysis and the Visual Explain views were spread across `src/types`,
`src/utils` and three component directories, and were entangled with code
that runs EXPLAIN: `explainPlan.ts` held `isDataModifyingQuery` (a question
about a statement, not a plan) and imported a type from a component, which
pointed the dependency from logic back at rendering.

Move all of it to a `@tabularis/explain` workspace package with three entry
points:

- `.` — types, metrics, diagnostics, stats, formatters, tree helpers. No
  runtime dependencies at all, so plan analysis runs in a browser, a worker
  or a plain Node script.
- `./react` — the graph, table, diagram, stats, node-details and bar views.
- `./flow` — the ReactFlow/dagre adapter, split out of `plan.ts` so the
  analysis core no longer pulls a graph library.

`isDataModifyingQuery` moves to `utils/sql.ts` beside `isExplainableQuery`,
where the rest of the statement-level questions live, and its tests move
into `tests/utils/sql.test.ts`.

What stays in the app, deliberately: the AI plan explanation (a host feature
on a host-configured provider), the Monaco raw-output tab (host theme
system), and the `VisualExplainView` composition that wires those two around
the package's components. Those host pieces are why the view was not portable
before.

Workspace consumers resolve the entry points to TypeScript source, so no
build step is needed for `dev`, `build`, `typecheck` or `vitest`;
`publishConfig` swaps in the tsup `dist` for publishing.

Verified: 3286 tests pass, app and package typecheck clean, eslint clean,
`pnpm build` and the package's tsup build both succeed.
