+++ package.json
@@ -22,6 +22,8 @@
     "test:rust": "cd src-tauri && cargo test",
     "test:all": "pnpm test:rust && pnpm test:coverage",
     "tauri": "tauri",
+    "build:explain": "pnpm --filter @tabularis/explain build",
+    "typecheck:explain": "pnpm --filter @tabularis/explain typecheck",
     "build:plugin-api": "pnpm --filter @tabularis/plugin-api build",
     "check:plugin-api": "pnpm --filter @tabularis/plugin-api check:sync",
     "build:create-plugin": "pnpm --filter @tabularis/create-plugin build",
@@ -36,6 +38,7 @@
     "@tailwindcss/postcss": "^4.2.2",
     "@tanstack/react-table": "^8.21.3",
     "@tanstack/react-virtual": "^3.13.23",
+    "@tabularis/explain": "workspace:*",
     "@tauri-apps/api": "^2.10.1",
     "@tauri-apps/plugin-clipboard-manager": "~2.3.2",
     "@tauri-apps/plugin-dialog": "^2.6.0",
