+++ packages/explain/package.json
@@ -1,0 +1,89 @@
+{
+  "name": "@tabularis/explain",
+  "version": "0.1.0",
+  "description": "Analyse and visualise a database EXPLAIN plan. Takes a parsed plan; never runs a query.",
+  "license": "Apache-2.0",
+  "homepage": "https://github.com/TabularisDB/tabularis/tree/main/packages/explain",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/TabularisDB/tabularis.git",
+    "directory": "packages/explain"
+  },
+  "bugs": {
+    "url": "https://github.com/TabularisDB/tabularis/issues"
+  },
+  "keywords": [
+    "explain",
+    "query-plan",
+    "postgres",
+    "mysql",
+    "database",
+    "visualisation"
+  ],
+  "type": "module",
+  "sideEffects": false,
+  "//": "Entry points resolve to TypeScript source so workspace consumers need no build step. publishConfig swaps in the compiled dist for npm.",
+  "exports": {
+    ".": "./src/index.ts",
+    "./react": "./src/react.ts",
+    "./flow": "./src/flow.ts"
+  },
+  "files": [
+    "dist",
+    "README.md",
+    "LICENSE"
+  ],
+  "scripts": {
+    "build": "tsup",
+    "clean": "rm -rf dist",
+    "typecheck": "tsc --noEmit",
+    "prepublishOnly": "pnpm run clean && pnpm run build"
+  },
+  "publishConfig": {
+    "access": "public",
+    "main": "./dist/index.js",
+    "types": "./dist/index.d.ts",
+    "exports": {
+      ".": {
+        "types": "./dist/index.d.ts",
+        "import": "./dist/index.js"
+      },
+      "./react": {
+        "types": "./dist/react.d.ts",
+        "import": "./dist/react.js"
+      },
+      "./flow": {
+        "types": "./dist/flow.d.ts",
+        "import": "./dist/flow.js"
+      }
+    }
+  },
+  "peerDependencies": {
+    "@xyflow/react": "^12.0.0",
+    "clsx": "^2.0.0",
+    "dagre": "^0.8.5",
+    "lucide-react": ">=0.400.0",
+    "react": "^18.0.0 || ^19.0.0",
+    "react-i18next": ">=15.0.0"
+  },
+  "peerDependenciesMeta": {
+    "@xyflow/react": { "optional": true },
+    "dagre": { "optional": true },
+    "clsx": { "optional": true },
+    "lucide-react": { "optional": true },
+    "react": { "optional": true },
+    "react-i18next": { "optional": true }
+  },
+  "devDependencies": {
+    "@types/dagre": "^0.7.54",
+    "@types/react": "^19.2.14",
+    "@xyflow/react": "^12.10.2",
+    "clsx": "^2.1.1",
+    "dagre": "^0.8.5",
+    "lucide-react": "^0.563.0",
+    "react": "^19.2.4",
+    "react-i18next": "^16.6.6",
+    "tsup": "^8.3.5",
+    "typescript": "~5.9.3"
+  }
+}
