+++ packages/explain/src/plan.ts
@@ -1,116 +1,11 @@
-import type { Node, Edge } from "@xyflow/react";
-import type { ExplainNode, ExplainPlan } from "../types/explain";
-import type { ExplainPlanNodeData } from "../components/ui/ExplainPlanNode";
-import type { ExplainMetrics } from "./explainMetrics";
-import { computeExplainMetrics } from "./explainMetrics";
-import type { ExplainDiagnostic } from "./explainDiagnostics";
-import dagre from "dagre";
-
-// ---------------------------------------------------------------------------
-// Tree → ReactFlow conversion
-// ---------------------------------------------------------------------------
-
 /**
- * Build the ReactFlow graph for a plan.
- *
- * `metrics` and `diagnostics` may be supplied by a caller that already computed
- * them for other views, so a plan is only walked once per render. Diagnostics
- * are passed in rather than derived here, which keeps this module independent of
- * `explainDiagnostics` — that module needs the helpers below.
+ * Plan analysis primitives: tree traversal, formatters, heat colours and the
+ * headline summary. No rendering library and no notion of the statement that
+ * produced the plan.
  */
-export function explainPlanToFlow(
-  plan: ExplainPlan,
-  selectedNodeId?: string | null,
-  metrics?: ExplainMetrics,
-  diagnostics?: Map<string, ExplainDiagnostic[]>,
-): {
-  nodes: Node[];
-  edges: Edge[];
-} {
-  const planMetrics = metrics ?? computeExplainMetrics(plan);
-  const rawNodes: Node[] = [];
-  const edges: Edge[] = [];
-
-  function walk(node: ExplainNode) {
-    const nodeMetrics = planMetrics.byId.get(node.id);
-    const data: ExplainPlanNodeData = {
-      node,
-      metrics: nodeMetrics ?? null,
-      maxExclusiveCost: planMetrics.maxExclusiveCost,
-      maxExclusiveTimeMs: planMetrics.maxExclusiveTimeMs,
-      diagnostics: diagnostics?.get(node.id) ?? [],
-      hasAnalyzeData: plan.has_analyze_data,
-      isSelected: selectedNodeId === node.id,
-    };
-
-    rawNodes.push({
-      id: node.id,
-      type: "explainPlan",
-      position: { x: 0, y: 0 },
-      data,
-    });
-
-    for (const child of node.children) {
-      edges.push({
-        id: `${node.id}-${child.id}`,
-        source: node.id,
-        target: child.id,
-        animated: true,
-        style: { stroke: "#6366f1" },
-      });
-      walk(child);
-    }
-  }
-
-  walk(plan.root);
-
-  return layoutExplainNodes(rawNodes, edges);
-}
-
-// ---------------------------------------------------------------------------
-// Dagre layout
-// ---------------------------------------------------------------------------
-
-export function layoutExplainNodes(
-  nodes: Node[],
-  edges: Edge[],
-): { nodes: Node[]; edges: Edge[] } {
-  const g = new dagre.graphlib.Graph();
-  g.setDefaultEdgeLabel(() => ({}));
-  g.setGraph({ rankdir: "TB", ranksep: 80, nodesep: 40 });
-
-  const NODE_WIDTH = 280;
-
-  for (const node of nodes) {
-    const data = node.data as ExplainPlanNodeData;
-    const lines =
-      3 +
-      (data.hasAnalyzeData ? 2 : 0) +
-      (data.node.filter ? 1 : 0) +
-      (data.diagnostics.length > 0 ? 1 : 0);
-    const height = 28 + lines * 22;
-    g.setNode(node.id, { width: NODE_WIDTH, height });
-  }
-
-  for (const edge of edges) {
-    g.setEdge(edge.source, edge.target);
-  }
-
-  dagre.layout(g);
-
-  const layoutedNodes = nodes.map((node) => {
-    const pos = g.node(node.id);
-    return {
-      ...node,
-      position: {
-        x: pos.x - NODE_WIDTH / 2,
-        y: pos.y - pos.height / 2,
-      },
-    };
-  });
-
-  return { nodes: layoutedNodes, edges };
-}
+import type { ExplainNode, ExplainPlan } from "./types";
+import type { ExplainMetrics } from "./metrics";
+import { computeExplainMetrics } from "./metrics";
 
 // ---------------------------------------------------------------------------
 // Color helpers
@@ -417,20 +312,4 @@ export function getExplainDriverLegend(plan: ExplainPlan): string[] {
     default:
       return [];
   }
-}
-
-// ---------------------------------------------------------------------------
-// Query type detection
-// ---------------------------------------------------------------------------
-
-export function isDataModifyingQuery(query: string): boolean {
-  const trimmed = query.trim().toUpperCase();
-  return (
-    trimmed.startsWith("INSERT") ||
-    trimmed.startsWith("UPDATE") ||
-    trimmed.startsWith("DELETE") ||
-    trimmed.startsWith("DROP") ||
-    trimmed.startsWith("ALTER") ||
-    trimmed.startsWith("TRUNCATE")
-  );
 }
