+++ packages/explain/tests/diagnostics.test.ts
@@ -5,9 +5,9 @@ import {
   getPlanDiagnostics,
   getWorstSeverity,
   type ExplainDiagnosticKind,
-} from "../../src/utils/explainDiagnostics";
-import { computeExplainMetrics } from "../../src/utils/explainMetrics";
-import type { ExplainNode, ExplainPlan } from "../../src/types/explain";
+} from "../src/diagnostics";
+import { computeExplainMetrics } from "../src/metrics";
+import type { ExplainNode, ExplainPlan } from "../src/types";
 
 function makeNode(overrides: Partial<ExplainNode> = {}): ExplainNode {
   return {
