+++ packages/explain/tests/plan.test.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect } from "vitest";
 import {
-  explainPlanToFlow,
   findExplainNode,
   flattenExplainNodes,
   getNodeCostStyle,
@@ -14,10 +13,8 @@ import {
   getMaxCost,
   getMaxTime,
   getRowEstimateRatio,
-  isDataModifyingQuery,
-} from "../../src/utils/explainPlan";
-import type { ExplainDiagnostic } from "../../src/utils/explainDiagnostics";
-import type { ExplainNode, ExplainPlan } from "../../src/types/explain";
+} from "../src/plan";
+import type { ExplainNode, ExplainPlan } from "../src/types";
 
 // ---------------------------------------------------------------------------
 // Helpers
@@ -59,112 +56,7 @@ function makePlan(overrides: Partial<ExplainPlan> = {}): ExplainPlan {
   };
 }
 
-// ---------------------------------------------------------------------------
-// explainPlanToFlow
-// ---------------------------------------------------------------------------
-
-describe("explainPlan", () => {
-  describe("explainPlanToFlow", () => {
-    it("should convert single-node plan to one ReactFlow node", () => {
-      const plan = makePlan();
-      const { nodes, edges } = explainPlanToFlow(plan);
-      expect(nodes).toHaveLength(1);
-      expect(edges).toHaveLength(0);
-      expect(nodes[0].id).toBe("node_0");
-      expect(nodes[0].type).toBe("explainPlan");
-    });
-
-    it("should create edges from parent to children", () => {
-      const child = makeNode({ id: "node_1", node_type: "Index Scan" });
-      const root = makeNode({ id: "node_0", children: [child] });
-      const plan = makePlan({ root });
-
-      const { nodes, edges } = explainPlanToFlow(plan);
-      expect(nodes).toHaveLength(2);
-      expect(edges).toHaveLength(1);
-      expect(edges[0].source).toBe("node_0");
-      expect(edges[0].target).toBe("node_1");
-    });
-
-    it("should handle deeply nested plans", () => {
-      const grandchild = makeNode({ id: "node_2" });
-      const child = makeNode({ id: "node_1", children: [grandchild] });
-      const root = makeNode({ id: "node_0", children: [child] });
-      const plan = makePlan({ root });
-
-      const { nodes, edges } = explainPlanToFlow(plan);
-      expect(nodes).toHaveLength(3);
-      expect(edges).toHaveLength(2);
-    });
-
-    it("should handle multiple children on same node", () => {
-      const child1 = makeNode({ id: "node_1" });
-      const child2 = makeNode({ id: "node_2" });
-      const root = makeNode({ id: "node_0", children: [child1, child2] });
-      const plan = makePlan({ root });
-
-      const { nodes, edges } = explainPlanToFlow(plan);
-      expect(nodes).toHaveLength(3);
-      expect(edges).toHaveLength(2);
-    });
-
-    it("should assign positions to all nodes", () => {
-      const child = makeNode({ id: "node_1" });
-      const root = makeNode({ id: "node_0", children: [child] });
-      const plan = makePlan({ root });
-
-      const { nodes } = explainPlanToFlow(plan);
-      for (const node of nodes) {
-        expect(node.position).toBeDefined();
-        expect(typeof node.position.x).toBe("number");
-        expect(typeof node.position.y).toBe("number");
-      }
-    });
-
-    it("should mark the selected node in data", () => {
-      const child = makeNode({ id: "node_1" });
-      const root = makeNode({ id: "node_0", children: [child] });
-      const plan = makePlan({ root });
-
-      const { nodes } = explainPlanToFlow(plan, "node_1");
-      const selectedNode = nodes.find((node) => node.id === "node_1");
-      const unselectedNode = nodes.find((node) => node.id === "node_0");
-
-      expect(selectedNode?.data.isSelected).toBe(true);
-      expect(unselectedNode?.data.isSelected).toBe(false);
-    });
-
-    it("should take diagnostics from the supplied map", () => {
-      const child = makeNode({ id: "node_1" });
-      const root = makeNode({ id: "node_0", children: [child] });
-      const plan = makePlan({ root });
-      const diagnostic: ExplainDiagnostic = {
-        kind: "hotspot",
-        severity: "critical",
-        labelKey: "editor.visualExplain.diagnostics.hotspot.label",
-        descriptionKey: "editor.visualExplain.diagnostics.hotspot.description",
-      };
-
-      const { nodes } = explainPlanToFlow(
-        plan,
-        null,
-        undefined,
-        new Map([["node_1", [diagnostic]]]),
-      );
-
-      expect(nodes.find((node) => node.id === "node_1")?.data.diagnostics).toEqual([
-        diagnostic,
-      ]);
-      expect(nodes.find((node) => node.id === "node_0")?.data.diagnostics).toEqual(
-        [],
-      );
-    });
-  });
-
-  // ---------------------------------------------------------------------------
-  // getNodeCostColor
-  // ---------------------------------------------------------------------------
-
+describe("plan", () => {
   describe("getNodeCostStyle", () => {
     it("should return green for low cost", () => {
       const style = getNodeCostStyle(10, 100);
@@ -454,52 +346,6 @@ describe("explainPlan", () => {
     it("should return sqlite notes", () => {
       const legend = getExplainDriverLegend(makePlan({ driver: "sqlite" }));
       expect(legend[0]).toBe("editor.visualExplain.sqliteLegend1");
-    });
-  });
-
-  // ---------------------------------------------------------------------------
-  // isDataModifyingQuery
-  // ---------------------------------------------------------------------------
-
-  describe("isDataModifyingQuery", () => {
-    it("should detect INSERT", () => {
-      expect(isDataModifyingQuery("INSERT INTO t VALUES (1)")).toBe(true);
-    });
-
-    it("should detect UPDATE", () => {
-      expect(isDataModifyingQuery("UPDATE t SET a = 1")).toBe(true);
-    });
-
-    it("should detect DELETE", () => {
-      expect(isDataModifyingQuery("DELETE FROM t WHERE id = 1")).toBe(true);
-    });
-
-    it("should detect DROP", () => {
-      expect(isDataModifyingQuery("DROP TABLE t")).toBe(true);
-    });
-
-    it("should detect ALTER", () => {
-      expect(isDataModifyingQuery("ALTER TABLE t ADD col INT")).toBe(true);
-    });
-
-    it("should detect TRUNCATE", () => {
-      expect(isDataModifyingQuery("TRUNCATE TABLE t")).toBe(true);
-    });
-
-    it("should not flag SELECT", () => {
-      expect(isDataModifyingQuery("SELECT * FROM t")).toBe(false);
-    });
-
-    it("should be case-insensitive", () => {
-      expect(isDataModifyingQuery("insert into t values (1)")).toBe(true);
-    });
-
-    it("should handle leading whitespace", () => {
-      expect(isDataModifyingQuery("  DELETE FROM t")).toBe(true);
-    });
-
-    it("should not flag WITH queries", () => {
-      expect(isDataModifyingQuery("WITH cte AS (SELECT 1) SELECT * FROM cte")).toBe(false);
     });
   });
 });
